refactor: move initializing work to appswitch
